refactor: staccato/category title 구분
