+++ backend/src/main/java/com/staccato/category/repository/CategoryMemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import com.staccato.category.domain.CategoryMember;
-import com.staccato.category.domain.Title;
+import com.staccato.category.domain.CategoryTitle;
 import com.staccato.member.domain.Member;
 
 public interface CategoryMemberRepository extends JpaRepository<CategoryMember, Long> {
@@ -22,7 +22,7 @@ public interface CategoryMemberRepository extends JpaRepository<CategoryMember, 
             """)
     List<CategoryMember> findAllByMemberIdAndDate(@Param("memberId") long memberId, @Param("date") LocalDate date);
 
-    boolean existsByMemberAndCategoryTitle(Member member, Title title);
+    boolean existsByMemberAndCategoryTitle(Member member, CategoryTitle title);
 
     @Modifying
     @Query("DELETE FROM CategoryMember mm WHERE mm.category.id = :categoryId")
