+++ backend/src/main/java/com/staccato/staccato/domain/StaccatoTitle.java
@@ -1,4 +1,4 @@
-package com.staccato.category.domain;
+package com.staccato.staccato.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -12,26 +12,21 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Embeddable
 @EqualsAndHashCode
-public class Title {
-    public static final int COLUMN_MAX_LENGTH = 50;
-    private static final int TITLE_MAX_LENGTH = 30;
+public class StaccatoTitle {
+    private static final int MAX_LENGTH = 30;
 
-    @Column(nullable = false, length = COLUMN_MAX_LENGTH)
+    @Column(nullable = false)
     private String title;
 
-    public Title(String title) {
+    public StaccatoTitle(String title) {
         String trimmedTitle = title.trim();
         validateLength(trimmedTitle);
         this.title = trimmedTitle;
     }
 
     private void validateLength(String trimmedTitle) {
-        if (trimmedTitle.isEmpty() || trimmedTitle.length() > TITLE_MAX_LENGTH) {
-            throw new StaccatoException("제목은 공백 포함 30자 이하로 설정해주세요.");
+        if (trimmedTitle.isEmpty() || trimmedTitle.length() > MAX_LENGTH) {
+            throw new StaccatoException("스타카토 제목은 공백 포함 30자 이하로 설정해주세요.");
         }
-    }
-
-    public boolean isSame(Title title) {
-        return this.title.equals(title.getTitle());
     }
 }
