+++ backend/src/test/java/com/staccato/category/domain/CategoryTitleTest.java
@@ -1,4 +1,4 @@
-package com.staccato.staccato.domain;
+package com.staccato.category.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class TitleTest {
+class CategoryTitleTest {
     @DisplayName("정상적인 제목은 생성에 성공한다.")
     @Test
     void createValidTitle() {
@@ -18,7 +18,7 @@ class TitleTest {
         String value = "title";
 
         // when
-        Title title = new Title(value);
+        CategoryTitle title = new CategoryTitle(value);
 
         // then
         assertThat(title.getTitle()).isEqualTo(value);
@@ -28,11 +28,11 @@ class TitleTest {
     @Test
     void createTitleAfterTrim() {
         // given
-        String value = " 스타카토 ";
-        String trimmedValue = "스타카토";
+        String value = " 여행의 추억 ";
+        String trimmedValue = "여행의 추억";
 
         // when
-        Title title = new Title(value);
+        CategoryTitle title = new CategoryTitle(value);
 
         // then
         assertThat(title.getTitle()).isEqualTo(trimmedValue);
@@ -46,9 +46,9 @@ class TitleTest {
         String value = "가".repeat(count);
 
         // when & then
-        assertThatThrownBy(() -> new Title(value))
+        assertThatThrownBy(() -> new CategoryTitle(value))
                 .isInstanceOf(StaccatoException.class)
-                .hasMessageContaining("스타카토 제목은 공백 포함 30자 이하로 설정해주세요.");
+                .hasMessageContaining("제목은 공백 포함 30자 이하로 설정해주세요.");
     }
 
     @DisplayName("제목이 1자 이상, 30자 이하이면 생성할 수 있다.")
@@ -59,6 +59,28 @@ class TitleTest {
         String value = "가".repeat(count);
 
         // when & then
-        assertThatNoException().isThrownBy(() -> new Title(value));
+        assertThatNoException().isThrownBy(() -> new CategoryTitle(value));
+    }
+
+    @DisplayName("isSame 메서드는 제목이 동일한 경우 true를 반환한다.")
+    @Test
+    void isSameWhenTitleIsEqual() {
+        // given
+        CategoryTitle title1 = new CategoryTitle("스터디");
+        CategoryTitle title2 = new CategoryTitle("스터디");
+
+        // when & then
+        assertThat(title1.isSame(title2)).isTrue();
+    }
+
+    @DisplayName("isSame 메서드는 제목이 다른 경우 false를 반환한다.")
+    @Test
+    void isNotSameWhenTitleIsDifferent() {
+        // given
+        CategoryTitle title1 = new CategoryTitle("스터디");
+        CategoryTitle title2 = new CategoryTitle("여행");
+
+        // when & then
+        assertThat(title1.isSame(title2)).isFalse();
     }
 }
