+++ backend/src/test/java/com/staccato/staccato/domain/CategoryTitleTest.java
@@ -1,4 +1,4 @@
-package com.staccato.category.domain;
+package com.staccato.staccato.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class TitleTest {
+class StaccatoTitleTest {
     @DisplayName("정상적인 제목은 생성에 성공한다.")
     @Test
     void createValidTitle() {
@@ -18,7 +18,7 @@ class TitleTest {
         String value = "title";
 
         // when
-        Title title = new Title(value);
+        StaccatoTitle title = new StaccatoTitle(value);
 
         // then
         assertThat(title.getTitle()).isEqualTo(value);
@@ -28,11 +28,11 @@ class TitleTest {
     @Test
     void createTitleAfterTrim() {
         // given
-        String value = " 여행의 추억 ";
-        String trimmedValue = "여행의 추억";
+        String value = " 스타카토 ";
+        String trimmedValue = "스타카토";
 
         // when
-        Title title = new Title(value);
+        StaccatoTitle title = new StaccatoTitle(value);
 
         // then
         assertThat(title.getTitle()).isEqualTo(trimmedValue);
@@ -46,9 +46,9 @@ class TitleTest {
         String value = "가".repeat(count);
 
         // when & then
-        assertThatThrownBy(() -> new Title(value))
+        assertThatThrownBy(() -> new StaccatoTitle(value))
                 .isInstanceOf(StaccatoException.class)
-                .hasMessageContaining("제목은 공백 포함 30자 이하로 설정해주세요.");
+                .hasMessageContaining("스타카토 제목은 공백 포함 30자 이하로 설정해주세요.");
     }
 
     @DisplayName("제목이 1자 이상, 30자 이하이면 생성할 수 있다.")
@@ -59,28 +59,6 @@ class TitleTest {
         String value = "가".repeat(count);
 
         // when & then
-        assertThatNoException().isThrownBy(() -> new Title(value));
-    }
-
-    @DisplayName("isSame 메서드는 제목이 동일한 경우 true를 반환한다.")
-    @Test
-    void isSameWhenTitleIsEqual() {
-        // given
-        Title title1 = new Title("스터디");
-        Title title2 = new Title("스터디");
-
-        // when & then
-        assertThat(title1.isSame(title2)).isTrue();
-    }
-
-    @DisplayName("isSame 메서드는 제목이 다른 경우 false를 반환한다.")
-    @Test
-    void isNotSameWhenTitleIsDifferent() {
-        // given
-        Title title1 = new Title("스터디");
-        Title title2 = new Title("여행");
-
-        // when & then
-        assertThat(title1.isSame(title2)).isFalse();
+        assertThatNoException().isThrownBy(() -> new StaccatoTitle(value));
     }
 }
